pyproject.toml
[tool.poetry]
name = "sandol-meal-service"
version = "0.1.0"
description = ""
authors = ["Seokyoung_Hong <tjrdud0330@naver.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "~3.11"
fastapi = "~0.115.6"
uvicorn = "^0.33"
SQLAlchemy = "^2"
pandas = "^2.2"
openpyxl = "^3.1"
httpx = "^0.28"
fastapi-pagination = "^0.12.34"
python-dotenv = "^1.0.1"
aiosqlite = "^0.21.0"
alembic = "^1.15.1"
cachetools = "^5.5.2"

[tool.poetry.group.dev.dependencies]
ruff = "^0.9"
black = "^25.1.0"
mypy = "^1.15.0"
types-cachetools = "^5.5.0.20240820"
types-pytz = "^2025.1.0.20250204"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".git-rewrite",
    ".hg",
    ".ipynb_checkpoints",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".pyenv",
    ".pytest_cache",
    ".pytype",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    ".vscode",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "site-packages",
    "venv",
    "docs",
    "alembic"
]

line-length = 88
indent-width = 4

target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "W", "C", "D", "R", "B", "N", "S", "PL"]
ignore = ["E203", "E501", "D415", "D403"]

# Allow fix for all enabled rules (when `--fix`) is provided.
fixable = ["ALL"]
unfixable = []

# Allow unused variables when underscore-prefixed.
dummy-variable-rgx = "^(_|__)$"

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[tool.ruff.lint.pydocstyle]
convention = "google"
